feat: loading skeleton when user landing on result page and loading icon when user loading more audio. refactoring project structure
