+++ src/api/swr-api/useSearch/settings.ts
@@ -1,3 +1,3 @@
-const swrOptions = { revalidateOnFocus: false, revalidateFirstPage: false };
+const swrOptions = { revalidateOnFocus: false, revalidateFirstPage: false, shouldRetryOnError: false };
 
 export default swrOptions;
