--- src/components/main/main.tsx
@@ -1,15 +1,0 @@
-import AppInfo from "./app-info/app-info.tsx";
-import Search from "./search/search.tsx";
-
-function Main() {
-  return (
-    <main>
-      <section>
-        <AppInfo />
-        <Search />
-      </section>
-    </main>
-  )
-}
-
-export default Main;
