--- src/components/main/search/search.tsx
@@ -1,11 +1,0 @@
-import SearchInput from "../search-input/search-input.tsx";
-
-function Search() {
-  return (
-    <div className="mt-8">
-      <SearchInput />
-    </div>
-  );
-}
-
-export default Search;
