--- src/components/music-player/music-card/style.module.css
@@ -1,6 +1,0 @@
-.container {
-  height: 600px;
-  background-size: 100% 100%;
-  border: 1px solid black;
-  box-shadow: 0 0 0 10px rgba(0, 0, 0, 0.2);
-}
