+++ src/components/skeletons/style.module.css
@@ -1,0 +1,22 @@
+@keyframes shimmer-element {
+  from {
+    transform: translateX(-100%);
+  }
+  to {
+    transform: translateX(100%);
+  }
+}
+
+.bgPlayerGradient::after {
+  position: absolute;
+  content: "";
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-image: linear-gradient(90deg, rgba(39, 0, 93, .10) 0, rgba(226, 110, 229, .10) 40%, rgba(226, 110, 229, .05) 60%, rgba(39, 0, 93, .10) 80%);
+  background-size: contain;
+  transform: translateX(-100%);
+  animation: shimmer-element 2s infinite linear;
+  will-change: transform;
+}
