+++ src/const/routes.ts
@@ -1,4 +1,4 @@
 export const ROUTES = {
-  "MAIN": "/",
-  "SEARCH_RESULT": "/search-result",
+  "INDEX": "/",
+  "RESULT": "/result",
 }
