+++ src/pages/index-page/app-info/app-info.tsx
