+++ src/pages/index-page/main-page.tsx
@@ -1,0 +1,13 @@
+import AppInfo from "./app-info/app-info.tsx";
+import Search from "./search/search.tsx";
+
+function MainPage() {
+  return (
+    <section>
+      <AppInfo />
+      <Search />
+    </section>
+  )
+}
+
+export default MainPage;
