+++ src/pages/index-page/search/search.tsx
@@ -1,0 +1,11 @@
+import SearchInput from "@src/components/search-input/search-input.tsx";
+
+function Search() {
+  return (
+    <div className="mt-8">
+      <SearchInput />
+    </div>
+  );
+}
+
+export default Search;
