+++ src/pages/result-page/audio-card/audio-footer/audio-line/audio-line.tsx
