+++ src/pages/result-page/audio-card/audio-footer/audio-line/style.module.css
