+++ src/pages/result-page/audio-card/audio-footer/audio-time-info/audio-time-info.tsx
@@ -1,0 +1,13 @@
+import useZustandStore from "@zustand/zustandStore.ts";
+
+function AudioTimeInfo() {
+  const currentSongInfo = useZustandStore(state => state.currentSongInfo);
+
+  return (
+    <div>
+      <p className="text-sm">{Math.floor(currentSongInfo.time)}s/{Math.floor(currentSongInfo.duration)}s</p>
+    </div>
+  );
+}
+
+export default AudioTimeInfo;
