+++ src/pages/result-page/audio-card/style.module.css
@@ -1,0 +1,5 @@
+.container {
+  background-size: 100% 100%;
+  border: 1px solid black;
+  box-shadow: 0 0 0 10px rgba(0, 0, 0, 0.2);
+}
