+++ src/routes-loaders.ts
@@ -2,17 +2,19 @@ import { LoaderFunctionArgs, redirect } from "react-router-dom";
 
 import { ROUTES } from "./const/routes.ts";
 
-const loadersCol = {
-  [ROUTES.SEARCH_RESULT]: {
+const routerLoaders = {
+  [ROUTES.RESULT]: {
     loader: async ({ request }: LoaderFunctionArgs) => {
       const url = new URL(request.url);
       const searchParam = url.searchParams.get("q");
+
       if (!searchParam) {
-        return redirect(ROUTES.MAIN);
+        return redirect(ROUTES.INDEX);
       }
+
       return {};
     }
   }
 }
 
-export default loadersCol;
+export default routerLoaders;
